% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sex.R
\name{draw_sex}
\alias{draw_sex}
\title{Draw sex values randomly for a list of individuals}
\usage{
draw_sex(n)
}
\arguments{
\item{n}{The number of individuals.}
}
\value{
The length-\code{n} vector of integer sex assignments: \code{1L} corresponds to male, \code{2L} to female.
}
\description{
Each individual has their sex drawn between male and female with equal probability.
Sex is encoded numerically following the convention for plink FAM files (see below).
}
\examples{
draw_sex( 10 )

}
\seealso{
Plink FAM format reference:
\url{https://www.cog-genomics.org/plink/1.9/formats#fam}
}
