% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indexes_chr}
\alias{indexes_chr}
\title{Extract last index of every chromosome in a vector of chromosomes}
\usage{
indexes_chr(chrs)
}
\arguments{
\item{chrs}{Vector of integer chromosomes as it appears in a \code{bim} table.}
}
\value{
Vector of end indexes per chromosome.
The numerical value of the chromosome is its index in this vector.
Chromosomes that were not observed have \code{NA} values as their end indexes.
}
\description{
This function tests that chromosomes appear in their numerical order (dies with an error otherwise), and returns the last index of each chromosome.
}
\examples{
# the end of chr1 is at index 4, for chr2 it's 7, and for chr3 it's 9
indexes_chr( c(1,1,1,1,2,2,2,3,3) )

}
