% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomb_founders_with_descendants.R
\name{recomb_founder_blocks_inherited}
\alias{recomb_founder_blocks_inherited}
\title{Identify IBD blocks of founders that were inherited by at least one individual in this data}
\usage{
recomb_founder_blocks_inherited(inds)
}
\arguments{
\item{inds}{The tidy table of recombination block data inherited in individuals from founders, such as the return value of \code{\link[=tidy_recomb_map_inds]{tidy_recomb_map_inds()}}.
More broadly, it is a table with a row for every IBD block, and at least these columns: "anc" is the label of the founder individual, "chr" is the chromosome number, "start" and "end" are the range of the block.
Although this data usually also has columns "ind" and "parent" identifying the individual and parental haplotype that inherited the block, this information is not used by this function and is removed if present.}
}
\value{
A table in a similar format as the input (columns "anc", "chr", "start", and "end") where each row is a contiguous IBD block whose each basepair is inherited by at least one individual.
}
\description{
Transforms a tidy table describing the founder blocks inherited by each individual into a more summarized table describing the contiguous blocks of each founder inherited by at least one individual.
This is especially useful when the focal individuals are several generations removed from the founders so only a small fraction of founder chromosomes are inherited.
}
\examples{
# manually construct a toy sample input,
# with individuals marked for clarity although they are not required
# note first two individuals
library(tibble)
inds <- tibble(
    ind = letters[1:3],
    parent = c('pat', 'mat', 'mat'),
    chr = c(1, 1, 2),
    start = c(   1,  800,   5),
    end   = c(1000, 2000, 100),
    anc   = c('f1', 'f1', 'f2')
)

# the new table merges the first two rows,
# because they overlapped from the same ancestor,
# while the third row stays unchanged (after removing individual info)
founder_blocks <- recomb_founder_blocks_inherited( inds )

}
\seealso{
\code{\link[=tidy_recomb_map_inds]{tidy_recomb_map_inds()}}
}
