% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomb_init_founders.R
\name{recomb_init_founders}
\alias{recomb_init_founders}
\title{Initialize chromosome strctures for founders}
\usage{
recomb_init_founders(ids, lengs)
}
\arguments{
\item{ids}{The list of IDs to use for each individual}

\item{lengs}{The lengths of each chromosome in centi-Morgans (cM).
If this vector is named, the output inherits these chromosome names.}
}
\value{
A named list of diploid individuals, each of which is a list with two haploid individuals named "pat" and "mat", each of which is a list of chromosomes (inherits names of \code{lengs} if present), each of which is a tibble with a single row and two columns: "end" equals the chromosome length, and "anc" equals the ID of the individual (from \code{ids}) concatenated with either "_pat" or "_mat" depending on which parent it is.
}
\description{
This function initializes what is otherwise a tedious structure for founders, to be used for simulating recombination in a pedigree.
The genetic structure is trivial, in that these "founder" chromosomes are each of a single ancestral individual (none are recombined).
}
\examples{
ancs <- recomb_init_founders( c('a', 'b'), c(100, 200) )
ancs

}
\seealso{
\code{\link[=recomb_fam]{recomb_fam()}} to simulate recombination across a pedigree using the founders initialized here.
}
